Refactor out to separate file
